requirement.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
django-service-objects==0.7.0
django-summernote==0.8.11.6
djangorestframework==3.12.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
